Move properties.hpp and change cmake
