+++ src/plugins/intel_npu/src/compiler/CMakeLists.txt
@@ -23,12 +23,6 @@ ov_add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
 target_include_directories(${TARGET_NAME}
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<BUILD_INTERFACE:${NPU_PLUGIN_SOURCE_DIR}/include>
-)
-target_include_directories(${TARGET_NAME}
-    SYSTEM PUBLIC
-        $<BUILD_INTERFACE:${NPU_PLUGIN_SOURCE_DIR}/thirdparty/level-zero/include>
-        $<BUILD_INTERFACE:${NPU_PLUGIN_SOURCE_DIR}/thirdparty/level-zero-ext>
 )
 
 target_link_libraries(${TARGET_NAME}
