+++ src/plugins/intel_npu/src/plugin/CMakeLists.txt
@@ -60,7 +60,6 @@ target_link_libraries(${TARGET_NAME}
 
 target_include_directories(${TARGET_NAME}
     PUBLIC
-        $<BUILD_INTERFACE:${NPU_PLUGIN_SOURCE_DIR}/include>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
 )
 target_include_directories(${TARGET_NAME}
